Added Game Review Functionality
Requires a peer review -- But this is basically fully complete with validation in the front end for the form, and validation in the javascript to send the user to an error page. The error pages need styling and the success page needs styling too.
This may not be able to be pulled to main because it is semi-dependant on other team member's tasks but if it comes to a point where it is required, then I will do so.