unfinished
Left to do:
Need to create two CRUD options in the api. 1st is to search for a review by gameID and userID to be used in javascript. 2nd is to update said review.

Final part of the js that need to be done to find the review by GameID and UserID to get the ReviewID, show it to user for editing then submitting it.