Validation modifications
All validation should be done within the actual files that use the fetch if the success page is done. My validation within deleteuser.js and resetuserpass.js perfectly displays this as the success pages cause errors but this could be mitigated using other methods, such as updating the current page itself however; this may not be as effective as you may think due to the possibility that the data will still send through to the API and without proper validation in the back-end, incorrect data may be uploaded, modified or deleted.