Improved Timeout System
Removed Cookies As Well
Timeout System:
Major development in making it work when closing the browser, refreshing the page, closing the tab or going onto a new page! It saves! This does not work across browsers, obviously, but if it is activated on another browser then it has the same functionality!